update admin api

+++ api/controller/Admin.php
@@ -18,10 +18,12 @@ if (strpos($url, "/") !== 0) {
 if ($url == '/admins' && $_SERVER['REQUEST_METHOD'] == 'GET') {
     $admins = $ADMIN->getAll($connect);
 
+    http_response_code(200);
     echo json_encode([
         "data" => $admins,
         "status" => "200",
-        "message" => "ok"
+        "message" => "ok",
+        "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
     ]);
 }
 
@@ -33,14 +35,19 @@ if (preg_match("/admins\/(\d+)/", $url, $matches) && $_SERVER['REQUEST_METHOD'] 
     $admin = $ADMIN->get($connect, $adminId);
 
     if ($admin) {
+        http_response_code(200);
         echo json_encode([
             "data" => $admin,
-            "message" => "ok"
+            "status" => "200",
+            "message" => "ok",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
         ]);
-    }
-    else {
+    } else {
+        http_response_code(404);
         echo json_encode([
-            "message" => "Admin not found!!"
+            "status" => "404",
+            "message" => "Admin not found!!",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
         ]);
     }
 }
@@ -50,20 +57,31 @@ if (preg_match("/admins\/(\d+)/", $url, $matches) && $_SERVER['REQUEST_METHOD'] 
  */
 if ($url == "/admins" &&  $_SERVER['REQUEST_METHOD'] == 'POST') {
     $input = json_decode(file_get_contents('php://input'), true);
+
+    $admin = $ADMIN->findByUser($connect, "username='{$input['username']}'");
+    if ($admin) {
+        http_response_code(409);
+        echo json_encode([
+            "status" => "409",
+            "message" => "Admin already exist!!",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
+        ]);
+        exit();
+    }
+
+    $input['password'] = password_hash($input['password'], PASSWORD_DEFAULT);
     $adminId = $ADMIN->create($connect, $input);
 
     if ($adminId) {
         $input['id'] = $adminId;
         $input['link'] = "/admins/$adminId";
 
+        http_response_code(201);
         echo json_encode([
             "data" => $input,
-            "message" => "ok"
-        ]);
-    }
-    else {
-        echo json_encode([
-            "message" => "Admin already exist!!"
+            "status" => "201",
+            "message" => "created",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
         ]);
     }
 }
@@ -75,10 +93,36 @@ if (preg_match("/users\/(\d+)/", $url, $matches) && $_SERVER['REQUEST_METHOD'] =
     $adminId = $matches[1];
     $input = json_decode(file_get_contents("php://input"), true);
 
-    $ADMIN->update($connect, $adminId, $input);
+    $admin = $ADMIN->get($connect, $adminId);
+    if (!$admin) {
+        http_response_code(404);
+        echo json_encode([
+            "status" => "404",
+            "message" => "Admin not found!!",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
+        ]);
+        exit();
+    }
 
+    if (!empty($input['username'])) {
+        $admin = $ADMIN->findByUser($connect, "username='{$input['username']}'");
+        if ($admin) {
+            http_response_code(409);
+            echo json_encode([
+                "status" => "409",
+                "message" => "User already exist!!",
+                "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
+            ]);
+            exit();
+        }
+    }
+
+    $ADMIN->update($connect, $adminId, $input);
+    http_response_code(200);
     echo json_encode([
-        "message" => "ok"
+        "status" => "200",
+        "message" => "ok",
+        "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
     ]);
 }
 
@@ -87,11 +131,24 @@ if (preg_match("/users\/(\d+)/", $url, $matches) && $_SERVER['REQUEST_METHOD'] =
  */
 if (preg_match("/admins\/(\d+)/", $url, $matches) && $_SERVER['REQUEST_METHOD'] == 'DELETE') {
     $adminId = $matches[1];
-    $ADMIN->delete($connect, $adminId);
 
-    echo json_encode([
-        "message" => "ok"
-    ]);
+    $admin = $ADMIN->get($connect, $adminId);
+    if (!$admin) {
+        http_response_code(404);
+        echo json_encode([
+            "status" => "404",
+            "message" => "Admin not found!!",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
+        ]);
+    } else {
+        $ADMIN->delete($connect, $adminId);
+        http_response_code(200);
+        echo json_encode([
+            "status" => "200",
+            "message" => "ok",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
+        ]);
+    }
 }
 
 /**
@@ -102,14 +159,20 @@ if ($url == "/admins/sign_in" &&  $_SERVER['REQUEST_METHOD'] == 'POST') {
     $admin = $ADMIN->signIn($connect, $input);
 
     if ($admin) {
+        http_response_code(200);
         echo json_encode([
             "data" => $admin,
-            "message" => "ok"
+            "status" => "200",
+            "message" => "ok",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
         ]);
     }
     else {
+        http_response_code(401);
         echo json_encode([
-            "message" => "Wrong username or password!!"
+            "status" => "401",
+            "message" => "Wrong username or password!!",
+            "time" => microtime(true) - $_SERVER["REQUEST_TIME_FLOAT"]
         ]);
     }
 }
